refactor UserCard and Content components of User pages
