--- src/pages/Users/components/Content/Content.stories.tsx
@@ -1,18 +1,0 @@
-import Content, { IContentProps } from ".";
-
-import { Story, Meta } from "@storybook/react";
-
-export default {
-    title: "Components/UserCard/Content",
-    component: Content,
-} as Meta;
-
-const Template: Story<IContentProps> = (args) => {
-    return <Content {...args} />;
-};
-
-export const Normal = Template.bind({});
-Normal.args = {
-    email: "binhtt@gmail.com",
-    name: "Thanh Bình",
-};
